+++ YoutubeDownloader/ViewModels/Dialogs/SettingsViewModel.cs
@@ -26,6 +26,11 @@ namespace YoutubeDownloader.ViewModels.Dialogs
             set => _settingsService.ShouldInjectTags = value;
         }
 
+        public bool ShouldSkipExistingFiles
+        {
+            get => _settingsService.ShouldSkipExistingFiles;
+            set => _settingsService.ShouldSkipExistingFiles = value;
+        }
         public string DefaultDownloadDirectory
         {
             get => _settingsService.DefaultDownloadDirectory;